Migrate regressions from 5501 to 7000
